requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --universal --python-version=3.12 pyproject.toml --group=dev -o requirements.txt
cfgv==3.4.0
    # via pre-commit
click==8.2.1
    # via typer
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
commonmark==0.9.1
    # via rich
distlib==0.4.0
    # via virtualenv
filelock==3.18.0
    # via virtualenv
identify==2.6.12
    # via pre-commit
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via pyodide-pack
markupsafe==3.0.2
    # via jinja2
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via pytest
platformdirs==4.3.8
    # via virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.2.0
    # via sql-bsky (pyproject.toml:dev)
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyodide-cli==0.3.0
    # via pyodide-pack
pyodide-pack==0.2.0
    # via sql-bsky (pyproject.toml:dev)
pytest==8.4.1
    # via sql-bsky (pyproject.toml:dev)
pyyaml==6.0.2
    # via pre-commit
rich==12.3.0
    # via
    #   pyodide-cli
    #   pyodide-pack
ruff==0.12.8
    # via sql-bsky (pyproject.toml:dev)
typer==0.4.2
    # via
    #   pyodide-cli
    #   pyodide-pack
virtualenv==20.33.1
    # via pre-commit
